Extracted common code to allow for reuse.
