+++ lib/screens/v2/candidate_generator/bus_shelter.ex
@@ -2,13 +2,14 @@ defmodule Screens.V2.CandidateGenerator.BusShelter do
   @moduledoc false
 
   alias Screens.Config.Screen
-  alias Screens.Config.V2.{BusShelter, EvergreenContentItem, Footer, Survey}
+  alias Screens.Config.V2.{BusShelter, Footer, Survey}
   alias Screens.Config.V2.Header.CurrentStopId
   alias Screens.V2.CandidateGenerator
   alias Screens.V2.CandidateGenerator.Widgets
   alias Screens.V2.Template.Builder
+  alias Screens.Util.Assets
 
-  alias Screens.V2.WidgetInstance.{EvergreenContent, LinkFooter, NormalHeader, SubwayStatus}
+  alias Screens.V2.WidgetInstance.{LinkFooter, NormalHeader, SubwayStatus}
 
   alias Screens.V2.WidgetInstance.Survey, as: SurveyInstance
 
@@ -50,7 +51,8 @@ defmodule Screens.V2.CandidateGenerator.BusShelter do
         now \\ DateTime.utc_now(),
         fetch_stop_name_fn \\ &fetch_stop_name/1,
         departures_instances_fn \\ &Widgets.Departures.departures_instances/1,
-        alert_instances_fn \\ &Widgets.Alerts.alert_instances/1
+        alert_instances_fn \\ &Widgets.Alerts.alert_instances/1,
+        evergreen_content_instances_fn \\ &Widgets.Evergreen.evergreen_content_instances/1
       ) do
     [
       fn -> header_instances(config, now, fetch_stop_name_fn) end,
@@ -58,7 +60,7 @@ defmodule Screens.V2.CandidateGenerator.BusShelter do
       fn -> alert_instances_fn.(config) end,
       fn -> footer_instances(config) end,
       fn -> subway_status_instances(config) end,
-      fn -> evergreen_content_instances(config) end,
+      fn -> evergreen_content_instances_fn.(config) end,
       fn -> survey_instances(config) end
     ]
     |> Task.async_stream(& &1.(), ordered: false, timeout: :infinity)
@@ -88,12 +90,6 @@ defmodule Screens.V2.CandidateGenerator.BusShelter do
     [%LinkFooter{screen: config, text: "More at", url: "mbta.com/stops/#{stop_id}"}]
   end
 
-  defp evergreen_content_instances(config) do
-    %Screen{app_params: %BusShelter{evergreen_content: evergreen_content}} = config
-
-    Enum.map(evergreen_content, &evergreen_content_instance(&1, config))
-  end
-
   defp survey_instances(config) do
     %Screen{
       app_params: %BusShelter{
@@ -109,31 +105,10 @@ defmodule Screens.V2.CandidateGenerator.BusShelter do
       %SurveyInstance{
         screen: config,
         enabled?: enabled,
-        medium_asset_url: s3_asset_url(medium_asset_path),
-        large_asset_url: s3_asset_url(large_asset_path)
+        medium_asset_url: Assets.s3_asset_url(medium_asset_path),
+        large_asset_url: Assets.s3_asset_url(large_asset_path)
       }
     ]
-  end
-
-  defp evergreen_content_instance(
-         %EvergreenContentItem{
-           slot_names: slot_names,
-           asset_path: asset_path,
-           priority: priority
-         },
-         config
-       ) do
-    %EvergreenContent{
-      screen: config,
-      slot_names: slot_names,
-      asset_url: s3_asset_url(asset_path),
-      priority: priority
-    }
-  end
-
-  defp s3_asset_url(asset_path) do
-    env = Application.get_env(:screens, :environment_name, "screens-prod")
-    "https://mbta-screens.s3.amazonaws.com/#{env}/#{asset_path}"
   end
 
   defp fetch_stop_name(stop_id) do
