+++ lib/screens/v2/candidate_generator/gl_eink.ex
@@ -2,7 +2,7 @@ defmodule Screens.V2.CandidateGenerator.GlEink do
   @moduledoc false
 
   alias Screens.Config.{Screen, V2}
-  alias Screens.Config.V2.{Footer, GlEink, Header, EvergreenContentItem}
+  alias Screens.Config.V2.{Footer, GlEink, Header}
   alias Screens.RoutePatterns.RoutePattern
   alias Screens.V2.CandidateGenerator
   alias Screens.V2.CandidateGenerator.Widgets
@@ -11,8 +11,7 @@ defmodule Screens.V2.CandidateGenerator.GlEink do
     FareInfoFooter,
     LineMap,
     NormalHeader,
-    Placeholder,
-    EvergreenContent
+    Placeholder
   }
 
   @scheduled_terminal_departure_lookback_seconds 180
@@ -46,7 +45,8 @@ defmodule Screens.V2.CandidateGenerator.GlEink do
         now \\ DateTime.utc_now(),
         fetch_destination_fn \\ &fetch_destination/2,
         departures_instances_fn \\ &Widgets.Departures.departures_instances/1,
-        alert_instances_fn \\ &Widgets.Alerts.alert_instances/1
+        alert_instances_fn \\ &Widgets.Alerts.alert_instances/1,
+        evergreen_content_instances_fn \\ &Widgets.Evergreen.evergreen_content_instances/1
       ) do
     [
       fn -> header_instances(config, now, fetch_destination_fn) end,
@@ -55,7 +55,7 @@ defmodule Screens.V2.CandidateGenerator.GlEink do
       fn -> footer_instances(config) end,
       fn -> placeholder_instances() end,
       fn -> line_map_instances(config) end,
-      fn -> evergreen_content_instances(config) end
+      fn -> evergreen_content_instances_fn.(config) end
     ]
     |> Task.async_stream(& &1.(), ordered: false, timeout: :infinity)
     |> Enum.flat_map(fn {:ok, instances} -> instances end)
@@ -135,32 +135,5 @@ defmodule Screens.V2.CandidateGenerator.GlEink do
       %Placeholder{color: :blue, slot_names: [:main_content]},
       %Placeholder{color: :green, slot_names: [:medium_flex]}
     ]
-  end
-
-  defp evergreen_content_instances(config) do
-    %Screen{app_params: %GlEink{evergreen_content: evergreen_content}} = config
-
-    Enum.map(evergreen_content, &evergreen_content_instance(&1, config))
-  end
-
-  defp evergreen_content_instance(
-         %EvergreenContentItem{
-           slot_names: slot_names,
-           asset_path: asset_path,
-           priority: priority
-         },
-         config
-       ) do
-    %EvergreenContent{
-      screen: config,
-      slot_names: slot_names,
-      asset_url: s3_asset_url(asset_path),
-      priority: priority
-    }
-  end
-
-  defp s3_asset_url(asset_path) do
-    env = Application.get_env(:screens, :environment_name, "screens-prod")
-    "https://mbta-screens.s3.amazonaws.com/#{env}/#{asset_path}"
   end
 end
